Fixed existing test and closed hole in comments
The tests needed to be rewritten because of the refactoring.

Comments testing on the new, separate comments controller are not yet
implemented.

A security hole was closed. When querying comments through the post
model, a filter could be applied to get user emails. This is not allowed
any longer for anonymous users. Further research needs to be done on
access that authenticated users have.